// front-end-app/quanly-thuysan/src/app/model/chebien.ts
export class CheBien {
  $class: string;
  idgdcb: string;
  sanphamnoi: string;
  sanphamngoai: string;
  nguyenlieu: string;
  cscb: string;
  soluongnhan: number;
  tgnhannl: string;
  masolonhan: string;
  tennguoinhan: string;
  khoiluongcattiet: string;
  khoiluongrua: string;
  tgrua: string;
  tennguoirua: string;
  nongdo: string;
  nhietdo: string;
  khoiluongsoche: string;
  tgbdsoche: string;
  tgktsoche: string;
  tennguoisoche: string;
  nhietdocd: string;
  tgcapdong: string;
  tennguoicd: string;
  tennccbb: string;
  soluongbb: string;
  loaibb: string;
  tennguoibg: string;
  masanpham: string;
  nhietdobq: string;
  tgxuat: string;
  soluonggiao: string;
  masologiao: string;
  diadiemgiao: string;
  phuongtien: string;
  cspp: string;
  tennguoivc: string;
  nhachebien: string;
}